+++ init_test.go
@@ -83,6 +83,16 @@ func TestUnexpectedInitReq(t *testing.T) {
 				errCode: ErrCodeProtocol,
 			},
 		},
+		{
+			name: "unexpected message type",
+			initMsg: &pingReq{
+				id: 1,
+			},
+			expectedError: errorMessage{
+				id:      1,
+				errCode: ErrCodeProtocol,
+			},
+		},
 	}
 
 	for _, tt := range tests {
@@ -178,7 +188,7 @@ func TestUnexpectedInitRes(t *testing.T) {
 				return
 			}
 
-			assert.Equal(t, ErrCodeProtocol, GetSystemErrorCode(err), "Unexpected error code")
+			assert.Equal(t, ErrCodeProtocol, GetSystemErrorCode(err), "Unexpected error code, got error: %v", err)
 			assert.Contains(t, err.Error(), tt.errMsg)
 		}()
 
@@ -316,14 +326,11 @@ func TestInitReqGetsError(t *testing.T) {
 	expectedErr := NewSystemError(ErrCodeBadRequest, "invalid host:port")
 	assert.Equal(t, expectedErr, err, "Error mismatch")
 	assert.Contains(t, logOut.String(),
-		"[E] Connection error.",
+		"[E] Failed during connection handshake.",
 		"Message should be logged")
 	assert.Contains(t, logOut.String(),
 		"tchannel error ErrCodeBadRequest: invalid host:port",
 		"Error should be logged")
-	assert.Contains(t, logOut.String(),
-		"site receive init res",
-		"Site should be logged")
 	close(connectionComplete)
 
 	<-listenerComplete