+++ peers/doc.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2015 Uber Technologies, Inc.
+// Copyright (c) 2017 Uber Technologies, Inc.
 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -18,9 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package tchannel
-
-// VersionInfo identifies the version of the TChannel library.
-// Due to lack of proper package management, this version string will
-// be maintained manually.
-const VersionInfo = "1.4.0"
+/*
+Package peers provides helpers for managing TChannel peers.
+*/
+package peers