+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+# v1.4.0
+
+* Add version information to the channel's LocalPeerInfo.
+* Add peers package for peer management utilities such as
+  consistent peer selection.
+* Fix SetScoreStrategy not rescoring existing peers. (#583).
+
 # v1.3.0
 
 * Exposes the channel's RootPeerList with `channel.RootPeers()`.